POSIX clocks documentation and minor fixes
git-svn-id: svn+ssh://svn.gna.org/svn/xenomai/trunk@676 c6d672ea-8702-0410-b560-f74c916a59fe

+++ ksrc/skins/posix/timer.c
@@ -246,8 +246,8 @@ int timer_settime (timer_t timerid,
     if ((unsigned) timerid >= PSE51_TIMER_MAX)
         goto einval;
 
-    if ((unsigned) value->it_value.tv_nsec >= ONE_BILLION ||
-        ((unsigned) value->it_interval.tv_nsec >= ONE_BILLION &&
+    if ((unsigned long) value->it_value.tv_nsec >= ONE_BILLION ||
+        ((unsigned long) value->it_interval.tv_nsec >= ONE_BILLION &&
          (value->it_value.tv_sec != 0 || value->it_value.tv_nsec != 0)))
         goto einval;
